requirements.txt
absl-py==2.1.0
accelerate @ git+https://github.com/huggingface/accelerate@b7fa2fa956f40e0b6f650d5eb1764680bf3fd8f7
aim==3.20.1
aim-ui==3.20.1
aimrecords==0.0.7
aimrocks==0.4.0
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-cors==0.7.0
aioitertools==0.11.0
aiorwlock==1.4.0
aiosignal==1.3.1
alembic==1.13.1
altair==5.3.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
art==6.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.15.0
base58==2.0.1
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blessed==1.20.0
blinker==1.8.2
bokeh==3.4.1
boto3==1.34.106
botocore==1.34.106
brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.29.3
colorcet==3.1.0
colorful==0.5.6
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cryptography==42.0.8
cycler==0.12.1
dask==2024.2.0
datasets==2.2.1
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.14.2
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distributed==2024.2.0
docker-pycreds==0.4.0
einops==0.8.0
et-xmlfile==1.1.0
evaluate==0.4.2
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.108.0
fastjsonschema==2.19.1
fastparquet==2024.5.0
filelock==3.14.0
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
gitpython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
googleapis-common-protos==1.63.1
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
hjson==3.1.0
holoviews==1.18.3
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub @ git+https://github.com/huggingface/huggingface_hub@919ce7d0ca281574a26cfa73cf242def95ac0119
hvplot==0.10.0
idna==3.7
imageio==2.34.1
importlib-metadata==7.1.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-client==8.6.2
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.1
jupyter-server-terminals==0.5.3
jupyterlab==4.2.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.2
jupyterlab-widgets==3.0.11
kiwisolver==1.4.5
linkify-it-py==2.0.3
lit==18.1.6
locket==1.0.0
lz4==4.3.3
mako==1.3.5
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mauve-text==0.3.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
notebook==7.2.0
notebook-shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.3.2
opencensus==0.11.4
opencensus-context==0.1.3
openpyxl==3.1.3
orjson==3.10.3
overrides==7.7.0
packaging==24.0
pandas==1.5.3
pandocfilters==1.5.1
panel==1.4.4
param==2.1.0
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
plotly==5.22.0
plotly-express==0.4.1
prometheus-client==0.20.0
prompt-toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==1.10.15
pydeck==0.9.1
pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyviz-comms==3.0.2
pyyaml==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
qtpy==2.4.1
ray==2.9.2
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
responses==0.18.0
restrictedpython==7.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3fs==2024.6.0
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
send2trash==1.8.3
sentence-transformers==3.0.0
sentencepiece==0.2.0
sentry-sdk==2.4.0
setproctitle==1.3.3
setuptools==70.0.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlalchemy==2.0.30
stack-data==0.6.3
starlette==0.32.0.post1
statsmodels==0.14.2
streamlit==1.35.0
sympy==1.12.1
tabulate==0.9.0
tblib==3.0.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
triton==2.0.0
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.1
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.1
uv==0.2.6
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
wandb==0.17.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xlrd==2.0.1
xlsxwriter==3.2.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.2